Removed test file test_csu_init and moved it to test_chainbuilder.py
